Move statement-specific config from SqlStatements back into StatementContext
