+++ core/src/test/java/org/jdbi/v3/core/statement/StatementContextTest.java
@@ -29,19 +29,17 @@ public class StatementContextTest {
     @Test(expected = IllegalArgumentException.class)
     public void testShouldNotBeAbleToCombineGeneratedKeysAndConcurrentUpdatable() throws Exception {
         final StatementContext context = StatementContextAccess.createContext();
-        final SqlStatements cfg = context.getConfig(SqlStatements.class);
 
-        cfg.setReturningGeneratedKeys(true);
-        cfg.setConcurrentUpdatable(true);
+        context.setReturningGeneratedKeys(true);
+        context.setConcurrentUpdatable(true);
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void testShouldNotBeAbleToCombineConcurrentUpdatableAndGeneratedKeys() throws Exception {
         final StatementContext context = StatementContextAccess.createContext();
-        final SqlStatements cfg = context.getConfig(SqlStatements.class);
 
-        cfg.setConcurrentUpdatable(true);
-        cfg.setReturningGeneratedKeys(true);
+        context.setConcurrentUpdatable(true);
+        context.setReturningGeneratedKeys(true);
     }
 
     private static class Foo {
